Change test order and split - there are limits to how much we can build and tear down in one go
