+++ tests/test_1000_poissonCart.py
@@ -12,11 +12,6 @@ unstructured_quad_box_regular_3D = uw.meshing.UnstructuredSimplexBox(
     minCoords=(0.0, 0.0, 0.0), maxCoords=(1.0, 1.0, 1.0), cellSize=0.1, regular=True
 )
 
-spherical_shell = uw.meshing.SphericalShell()
-annulus = uw.meshing.Annulus()
-cubed_sphere = uw.meshing.CubedSphere()
-
-
 # Maybe lower and upper would work better for the names of the box mesh boundaries too.
 
 
@@ -45,20 +40,5 @@ def test_poisson_boxmesh(mesh):
 
     assert poisson.snes.getConvergedReason() > 0
 
-
-@pytest.mark.parametrize("mesh", [spherical_shell, annulus, cubed_sphere])
-def test_poisson_sphere(mesh):
-    u = uw.discretisation.MeshVariable(
-        r"mathbf{u}", mesh, 1, vtype=uw.VarType.SCALAR, degree=2
-    )
-
-    poisson = uw.systems.Poisson(mesh, u_Field=u)
-    poisson.constitutive_model = uw.systems.constitutive_models.DiffusionModel(mesh.dim)
-    poisson.constitutive_model.Parameters.diffusivity = 1
-
-    poisson.f = 0.0
-    poisson.add_dirichlet_bc(1.0, "Lower")
-    poisson.add_dirichlet_bc(0.0, "Upper")
-    poisson.solve()
-
-    assert poisson.snes.getConvergedReason() > 0
+    del poisson
+    del mesh
